Fix capitalization issue for WellService.js
